+++ Changelog.md
@@ -31,6 +31,9 @@
 - Added the `Stockable` interface to the `Product` and `MasterProductVariant` models
 - Added the `backorder` field to products and product variants 
 - BC: Added the `?CheckoutSubject` return type to the `getCart()` method of the `Checkout` interface
+- BC: Changed `Checkout::getShippingAddress()` return type to be nullable
+- BC: Added the void return type to `Checkout::setShippingAddress()`
+- BC: Added the `removeShippingAddress()` method to the Checkout interface
 - BC: The unused `$config` parameter has been removed from the `RequestStore` checkout driver constructor
 - Fixed possible null return type on Billpayer::getName() when is_organization is true but the company name is null